--- src/main/java/com/yanolja/areas/auth/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package com.yanolja.areas.auth.repository;
-
-import com.yanolja.areas.auth.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email);
-
-}
