+++ src/main/java/com/yanolja/areas/user/domain/SocialProvider.java
@@ -1,4 +1,4 @@
-package com.yanolja.areas.auth.domain;
+package com.yanolja.areas.user.domain;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -6,13 +6,16 @@ import lombok.AllArgsConstructor;
 
 @AllArgsConstructor
 @JsonFormat(shape = JsonFormat.Shape.OBJECT)
-public enum UserRole {
-    @Schema(description = "유저 권한")
-    USER("일반 사용자"),
-    ADMIN("관리자");
+public enum SocialProvider {
+    @Schema(description = "구글 로그인")
+    GOOGLE("구글"),
+    
+    @Schema(description = "카카오 로그인")
+    KAKAO("카카오");
 
     private String label;
-    
+
+
     public String getLabel() {
         return this.label;
     }
