+++ src/main/java/com/yanolja/areas/user/domain/UserDetail.java
@@ -1,4 +1,4 @@
-package com.yanolja.areas.auth.domain;
+package com.yanolja.areas.user.domain;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
