+++ src/main/java/com/yanolja/areas/user/repository/UserRepository.java
@@ -1,0 +1,9 @@
+package com.yanolja.areas.user.repository;
+
+import com.yanolja.areas.user.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long>, UserRepositoryCustom {
+}
